D2: updating tests and inter-model logic.
When deleting cart: line_items get deleted as well

But, we can't delete a Product that has LineItems associated. For this we use a before_destroy_hook, that checks the existance of line_items in the Product being destroyed.

We also update the LineItem Fixture to be able to test this